% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaA.R
\name{metaA}
\alias{metaA}
\title{Metabolites analysis}
\usage{
metaA(
  df_meta,
  df_clinical,
  trt_var,
  clinical_var,
  adjustment_var = c(),
  no_signal_correction = TRUE,
  plot_volcano = TRUE,
  output_trt_association = TRUE,
  output_clin_rainplot = TRUE
)
}
\arguments{
\item{df_meta, }{experimental metabolites data}

\item{df_clinical, }{clinical data}

\item{trt_var, }{treatment variable}

\item{clinical_var, }{clinical variables in the rainplot}

\item{adjustment_var, }{adjustment variable in the linear model}

\item{no_signal_correction, }{defaults to true}

\item{plot_volcano, }{defaults to true}

\item{output_trt_association, }{defaults to true}

\item{output_clin_rainplot, }{defaults to true}
}
\value{
list showing linear model output
}
\description{
This function doing metabolites vs treatment association, metabolites vs clinical variable association.
It assumes that the first column of df_meta
contains the subjectId, with or without year, plate_well in the second and third column.  in the subsequent columns
are metabolites measurements. It assumes the first column of df_clinical contains subjectId for patients invovled
in the study, it then contains year if it has measurements in two years. trt_var is a character, showign the treatment variable, and clinical_var is a character vector showing
clinical variable used to plot the rainplot to show the clinical varialbe and metabolites association.
adjust_var shows the clinical variables used in the model for adjustment, it is a character vector.
}
