% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaQ2.R
\name{metaQ2}
\alias{metaQ2}
\title{Checking metabolites quality2}
\usage{
metaQ2(
  df_sample,
  df_pp,
  plot_missing = TRUE,
  output_missing = TRUE,
  plot_CV = TRUE,
  output_CV = TRUE,
  output_meta_summary = TRUE,
  plot_well_median = TRUE
)
}
\arguments{
\item{df_sample, }{experimental metabolites data}

\item{df_pp, }{pooled plasma data}

\item{plot_missing, }{defaults to true}

\item{output_missing, }{defaults to true}

\item{plot_CV, }{defaults to true}

\item{output_CV, }{defaults to true}

\item{output_meta_summary, }{defaults to true}

\item{plot_well_median, }{defaults to true}
}
\value{
list showing missing, cv, meta_summary
}
\description{
This function check the metabolites data quality. It assumes that the first column of df_sample
contains the subjectId, with or without year, plate_well in the second and third column.  in the subsequent columns
are metabolites measurements. It assumes the first column of df_pp contains plate_well position for the pooled plasma
samples.
}
